+++ CotEditor/Sources/EditorTextView.swift
@@ -70,6 +70,9 @@ final class EditorTextView: NSTextView, Themable {
     
     private var lineHighLightColor: NSColor?
     
+    private let instanceHighlightColor: NSColor = NSColor(calibratedHue: 0.24, saturation: 0.8, brightness: 0.8, alpha: 0.2)
+    private lazy var instanceHighlightTask = Debouncer(delay: .milliseconds(500)) { [unowned self] in self.highlightInstance() }
+    
     private var needsRecompletion = false
     private var particalCompletionWord: String?
     private lazy var completionTask = Debouncer(delay: .seconds(0)) { [unowned self] in self.performCompletion() }  // NSTextView cannot be weak
@@ -499,6 +502,10 @@ final class EditorTextView: NSTextView, Themable {
             let bracePairs = BracePair.braces + (UserDefaults.standard[.highlightLtGt] ? [.ltgt] : [])
             self.highligtMatchingBrace(candidates: bracePairs)
         }
+        
+        // invalidate current instances highlight
+        self.layoutManager?.removeTemporaryAttribute(.roundedBackgroundColor, forCharacterRange: self.string.nsRange)
+        self.instanceHighlightTask.schedule()
     }
     
     
@@ -622,6 +629,8 @@ final class EditorTextView: NSTextView, Themable {
             
             NSGraphicsContext.restoreGraphicsState()
         }
+        
+        self.drawRoundedBackground(in: rect)
     }
     
     
@@ -1258,6 +1267,29 @@ final class EditorTextView: NSTextView, Themable {
         self.didChangeText()
         
         return true
+    }
+    
+    
+    /// highlight all instances of the selection
+    private func highlightInstance() {
+        
+        guard
+            self.selectedRanges.count == 1,
+            self.selectedRange.length > 0,
+            (try! NSRegularExpression(pattern: "^\\b\\w.*\\w\\b$"))
+                .firstMatch(in: self.string, options: [.withTransparentBounds], range: self.selectedRange) != nil,
+            let range = Range(self.selectedRange, in: self.string)
+            else { return }
+        
+        let substring = String(self.string[range])
+        let pattern = "\\b" + NSRegularExpression.escapedPattern(for: substring) + "\\b"
+        let regex = try! NSRegularExpression(pattern: pattern)
+        
+        regex.matches(in: self.string, range: self.string.nsRange)
+            .map { $0.range }
+            .forEach {
+                self.layoutManager?.addTemporaryAttribute(.roundedBackgroundColor, value: self.instanceHighlightColor, forCharacterRange: $0)
+            }
     }
     
 }